requirements.txt
accelerate
hydra-core
torch
transformers
datasets
pandas
numpy
Pillow
torchmetrics
torchvision
wandb